Add MSSQL dialect and fix up the postgres' identifier rules
The `@@version` test is MS' dialect of SQL, it seems, so test it with
its own dialect.

Update the rules for identifiers in Postresql dialect per documentation,
while we're at it. The current identifier rules in Postgresql dialect
were introduced in this commit - as a copy of generic rules, it seems:
https://github.com/andygrove/sqlparser-rs/commit/810cd8e6cf28a1586579bbd5d7dfcbae078c2922#diff-2808df0fba0aed85f9d35c167bd6a5f1L138