Improve some literal oddness in docs
Square brackets, when run on nightly, auto-link to the types that they refer
to. That doesn't work if the types are meant to be a grammar, not a reference
to a type, and it gives crazy warnings.

Also it just looks a bit more reasonable IMO to have literal terms look like
literal terms.