provide parse_all_or_distinct abstraction
This consolidates the act of checking for both `ALL` and `DISTINCT` into
a parser function. Please note that this present implementation reduces
the information provided by the resulting error message. This may be
undesirable, so pending review, may certainly be updated or reverted.
However it may also be the case that the additional contextual
information is not necessary.