Move test_utils out of library
Previously the test_utils were part of the public API (albeit their
documentation was hidden), so that they could be used by a single unit
test. Since the unit test only used public APIs, we can just move it
to the integration tests, allowing us to outsource the test_utils.