Add a recursion limit to prevent stack overflows
Until now, it's been able to trigger a stack overflow crash by providing a
string with excessive recursion. For instance a string of 1000 left brackets
causes the parser to recurse down 1000 times, and overflow the stack.

This commit adds protection against excessive recursion. It adds a field to
`Parser` for tracking the current recursion depth. Every function that returns
a `Result` gains a recursion depth check. This isn't quite every method on the
`Parser`, but it's the vast majority.

An alternative implemention would be to only protect against AST recursions,
rather than recursive function calls in `Parser`. That isn't as easy to implement
because the parser is so large.