ISSUE-1088: Fix array_agg wildcard behavior
The behavior for qualified wildcards in array_agg is broken for
postgresql, it throws an error when it should parse. This commit
partially reverts 823635d2fcd2642254b14b5737c9758d96a9580b, and adds
some test coverage for this behavior.

Notably this moves the WildcardExpr enum back into the main Expr enum,
and adjusts usage/impl for it as needed.