Tokenize at signs separately from subsequent strings
This is needed for parsing MySQL-style `'user'@'host'` grantee syntax.

As far as I can tell, no dialect allows quotes or backticks as part of
an identifier (regardless of whether it starts with `@`) without other
specific syntax (e.g. nested in another quote style and thus not
starting with `@`), so this shouldn't adversely affect non-MySQL
dialects.