Reorganize the Parser module
I mostly did this as an exercise to get a general feel of how the Parser
implementation is organized. The basics here are that for every top
level keyword in Parser::parse_statement I created a new module and
moved the corresponding function to that module. Then I spent a few
hours checking `find references` and any method that was in a single new
module got moved there.

Towards the end I started making some arbitary decisions on where
functions referenced from multiple modules lived. Some of these seemed
obvious, while some were certainly arbitrary.

Most of the motivation here was that working on a 13,000 line file was
causing my editor to be very not happy. After this happy, the largest
module is now src/parser/select.rs which clocks in at 2142 lines.

I should note, that the only visible changes are hopefully a few
functions that had visibility flipped from private to public because I
forgot about pub(crate) when I first started. Other than that, this is
purely copy/paste moving of code to new module files.