Replace expect_keyword with expect_keyword_is
Except for a single instance, every use of expect_keyword was ignoring
the returned token. This adds a new `expect_keyword_is` that avoids that
unnecessary clone.

I nearly added a `#[must_use]` attribute to the `expect_keyword` method,
but decided against it as that feels like a breaking API change even if
it would nudge folks toward the correct method.