Fix: Snowflake ALTER SESSION cannot be followed by other statements.
Fixes https://github.com/apache/datafusion-sqlparser-rs/issues/1775.

Currently parse_session_options does not check for semicolons, which
makes it impossible to parse multiple statements, where a statement
comes after ALTER SESSION.

Furthermore, it seems we should not advance the parser onto the
semicolon as it is otherwise skipped in the next loop iteration
of parse_statements() where the parser advances to the next token,
which would then result in "Expected end of statement".