Add support for MSSQL IF/ELSE statements.
These are syntactically quite different from the already supported
IF ... THEN ... ELSEIF ... END IF statements. Hence IfStatement is now
an enum with two variants and statement parsing is overridden for
the MSSQL dialect in order to parse IF statements differently for MSSQL.

Thereby fix spans for if/case AST nodes by including start/end tokens,
if present.