Support MySQL index options such as `USING` in `CREATE INDEX`
For MySQL, the preferred position to put `USING BTREE` is after the
column list rather than before; before is still supported but considered
[deprecated]:

> The preferred position is after the column list. Expect support for
  use of the option before the column list to be removed in a future
  MySQL release.

[deprecated]: https://dev.mysql.com/doc/refman/8.4/en/alter-table.html

We already parse this correctly for `ALTER TABLE` statements to add
keys, and in table-level constraint parsing for `CREATE TABLE`. We can
just reuse the same parsing to support the newer style for `CREATE
INDEX` statements as well.